Changes to the page_list/thumbnail_grid template
The existing template was missing the date variable so fixed that.

If a page had no thumbnail it would just render an empty div. So conditionally adding thumbnail or not is now there and also depending on settings in page list you can show/hide thumbnails which, understandably defeats the purpose of a "thumbnail_list" to have items without thumbnails but at the very least if no picture it present it could output the title/desc/button. Better than just a empty div. So that is here. 

RSS feed was also missing so that is there too now. I wrapped the foreach $pages loop with another div as well because again if you were  trying to apply custom styles to this, it would be a problem, the title div and now the rss div would get in the way of doing things like :first-child/:nth-child in CSS/LESS. I guess it also will allow you to apply display: flex to that containing element for doing flexy stuff.

Button now renders if you choose to turn it on with custom text. I know there was an overlay hover effect on the image but it wasn't obvious to me at least that, that is what the control is doing in the page list settings. As follows, the $title will link or not link appropriately if button is used.

Also put in $columnClass variable and just put 'col-sm-4' in there as a bootstrap class to create the grid of items. It doesn't create a grid layout by default and I'm going to pull request the view.css for this template too because flexbox would be needed to stack things up neatly. Frankly I don't know how you best handle this because col-sm-4 is meaningless to someone using like Foundation in their theme or some other grid. It seems kind of taboo to apply any styling to vanilla blocks but in this way it would at least just be a meaningless class if bootstrap isn't being used in the theme, it won't stop it from rendering at the end of the day and they could still apply styles in a number of different ways from override folders or whatever.